chore: move db handle function to a new file
