+++ internal/database/dialer/postgres/client.go
@@ -381,148 +381,6 @@ func (c *client) SaveDatasetENSNamehash(ctx context.Context, namehash *model.ENS
 	return c.database.WithContext(ctx).Clauses(clauses...).Create(&value).Error
 }
 
-func (c *client) UpdateRecentMastodonHandles(ctx context.Context, handles []*model.MastodonHandle) error {
-	logger := zap.L()
-	logger.Info("Starting UpdateRecentMastodonHandles", zap.Int("handleCount", len(handles)))
-
-	// ToDo: remove hardcoded db table name and fix the function logic
-	tableName := "dataset_mastodon_update_handles"
-	// Construct the SQL query
-	sql, valueArgs := buildInsertQuery(handles, tableName)
-
-	logger.Debug("Generated SQL", zap.String("sql", sql), zap.Any("args", valueArgs))
-
-	// Execute the update query
-	result := c.database.WithContext(ctx).Exec(sql, valueArgs...)
-	if result.Error != nil {
-		logger.Error("Error updating recent handles", zap.Error(result.Error))
-		return fmt.Errorf("failed to update recent handles: %w", result.Error)
-	}
-
-	// Trim old handles
-	if err := trimOldHandles(ctx, c, tableName); err != nil {
-		logger.Error("Error trimming old handles", zap.Error(err))
-		return err
-	}
-
-	logger.Info("Completed UpdateRecentMastodonHandles",
-		zap.Int64("updatedHandles", result.RowsAffected))
-
-	return nil
-}
-
-// buildInsertQuery generates the SQL insert query and corresponding values for Mastodon handles
-func buildInsertQuery(handles []*model.MastodonHandle, tableName string) (string, []interface{}) {
-	valueStrings := make([]string, len(handles))
-	valueArgs := make([]interface{}, 0, len(handles)*2)
-
-	for i, h := range handles {
-		valueStrings[i] = fmt.Sprintf("($%d, $%d)", i*2+1, i*2+2)
-
-		valueArgs = append(valueArgs, h.Handle, h.LastUpdated)
-	}
-
-	// Insert SQL with ON CONFLICT clause
-	sql := fmt.Sprintf(`
-        INSERT INTO %s (handle, last_updated)
-        VALUES %s
-        ON CONFLICT (handle) DO UPDATE SET last_updated = EXCLUDED.last_updated;`,
-		tableName, strings.Join(valueStrings, ","),
-	)
-
-	return sql, valueArgs
-}
-
-// trimOldHandles removes outdated handles from the dataset
-func trimOldHandles(ctx context.Context, c *client, tableName string) error {
-	trimResult := c.database.WithContext(ctx).Exec(fmt.Sprintf(`
-        DELETE FROM %s
-        WHERE handle NOT IN (
-            SELECT handle FROM %s
-            ORDER BY last_updated DESC
-            LIMIT $1
-        )`, tableName, tableName), maxRecentHandles)
-
-	if trimResult.Error != nil {
-		return fmt.Errorf("failed to trim old handles: %w", trimResult.Error)
-	}
-
-	return nil
-}
-
-func (c *client) GetUpdatedMastodonHandles(ctx context.Context, since uint64, limit int, cursor string) (*model.PaginatedMastodonHandles, error) {
-	var result []model.MastodonHandle
-
-	var totalCount int64
-
-	// ToDo: remove hardcoded db table name
-	// Query for the handles table
-	query := c.database.WithContext(ctx).
-		Table("dataset_mastodon_update_handles")
-
-	// If 'since' is greater than 0, we're fetching updated handles; otherwise, fetch all handles
-	if since > 0 {
-		query = query.Where("last_updated > ?", since)
-	}
-
-	// Count total handles (or updated handles)
-	if err := query.Count(&totalCount).Error; err != nil {
-		return nil, fmt.Errorf("failed to count handles: %w", err)
-	}
-
-	// Apply cursor if provided
-	if cursor != "" {
-		var cursorLastUpdated uint64
-
-		var cursorHandle string
-
-		// Decode the cursor format: "<last_updated>:<handle>"
-		_, err := fmt.Sscanf(cursor, "%d:%s", &cursorLastUpdated, &cursorHandle)
-		if err != nil {
-			return nil, fmt.Errorf("invalid cursor format: %w", err)
-		}
-
-		query = query.Where("last_updated < ? OR (last_updated = ? AND handle > ?)", cursorLastUpdated, cursorLastUpdated, cursorHandle)
-	}
-
-	// Fetch handles with their last_updated timestamps
-	err := query.
-		Select("handle, last_updated").
-		Order("last_updated DESC, handle ASC").
-		Limit(limit + 1).
-		Find(&result).
-		Error
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to get handles: %w", err)
-	}
-
-	// Determine the next cursor if there are more than 'limit' handles
-	var nextCursor string
-
-	if len(result) > limit {
-		lastHandle := result[limit-1]
-
-		// Create the next cursor from the 'last_updated' and 'handle'
-		nextCursor = fmt.Sprintf("%d:%s", lastHandle.LastUpdated, lastHandle.Handle)
-
-		// Remove the extra handle for pagination
-		result = result[:limit]
-	}
-
-	// Extract just the handles for the response
-	handles := make([]string, len(result))
-	for i, r := range result {
-		handles[i] = r.Handle
-	}
-
-	return &model.PaginatedMastodonHandles{
-		Handles:    handles,
-		TotalCount: totalCount,
-		NextCursor: nextCursor,
-	}, nil
-}
-
 // Dial dials a database.
 func Dial(ctx context.Context, dataSourceName string, partition bool) (database.Client, error) {
 	var err error
